+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Fix transitive selector refresh for some cases (#1409)
 - Run atom effects when atoms are initialized from a set during a transaction from `useRecoilTransaction_UNSTABLE()` (#1466)
 - Unsubscribe `onSet()` handlers in atom effects when atoms are cleaned up. (#1509)
+- Atom effects are cleaned up when initialized by a Snapshot which is released. (#1511)
 - Avoid extra re-renders in some cases when a component uses a different atom/selector. (#825)
 - `<RecoilRoot>` will only call `initializeState()` once during the initial render. (#1372)